refactor(desktop): group the brain and voice main-process modules
