+++ apps/desktop/src/main/voice/credential-transition.ts
